[ui] show/hide docker, add as component
